Moved sigma container into a little wrapper. Moved helper functions to graph directory
